New custom-written GeoIP library. It's pretty slick.
There was already a similar library for node.js at <https://github.com/strange/node-geoip>, but it just calls the C API, so it's blocking. Plus the interface is kind of ugly (instead of passing a callback, you have to add a listener that isn't directly associated with your lookup).

The database tables will be removed from the schema later.